Updated based on code review
Moved MavenResource bits to the TaskBatchAutoConfigure.
Set Maven deployer dependencies to optional
other code cleanup
