+++ spring-cloud-task-batch/src/main/java/org/springframework/cloud/task/batch/configuration/TaskBatchAutoConfiguration.java
@@ -16,18 +16,26 @@
 package org.springframework.cloud.task.batch.configuration;
 
 import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
 
 import javax.sql.DataSource;
 
 import org.springframework.batch.core.Job;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.cloud.deployer.resource.maven.MavenProperties;
+import org.springframework.cloud.deployer.resource.maven.MavenResourceLoader;
+import org.springframework.cloud.deployer.resource.support.DelegatingResourceLoader;
 import org.springframework.cloud.task.batch.listener.TaskBatchExecutionListener;
 import org.springframework.cloud.task.repository.TaskExplorer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.io.ResourceLoader;
 import org.springframework.util.CollectionUtils;
 
 /**
@@ -64,6 +72,33 @@ public class TaskBatchAutoConfiguration {
 			else {
 				throw new IllegalStateException("Expected one datasource and found " + dataSources.size());
 			}
+		}
+	}
+
+	@Configuration
+	@ConditionalOnClass(name = "org.springframework.cloud.deployer.resource.maven.MavenResourceLoader")
+	public static class MavenResourceConfiguration {
+		@Bean
+		public MavenResourceLoader mavenResourceLoader(MavenProperties properties) {
+			return new MavenResourceLoader(properties);
+		}
+
+		@Bean
+		@ConditionalOnMissingBean(DelegatingResourceLoader.class)
+		public DelegatingResourceLoader delegatingResourceLoader(MavenResourceLoader mavenResourceLoader) {
+			Map<String, ResourceLoader> loaders = new HashMap<>();
+			loaders.put("maven", mavenResourceLoader);
+			return new DelegatingResourceLoader(loaders);
+		}
+
+		@Bean
+		public MavenProperties mavenProperties() {
+			return new MavenConfigurationProperties();
+		}
+
+		@ConfigurationProperties(prefix = "maven")
+		static class MavenConfigurationProperties extends MavenProperties {
+
 		}
 	}
 }
