+++ spring-cloud-task-samples/tasksink/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+maven.remoteRepositories.springRepo.url=https://repo.spring.io/libs-snapshot-local
