+++ spring-cloud-task-stream/pom.xml
@@ -70,6 +70,7 @@
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-deployer-resource-support</artifactId>
+			<optional>true</optional>
 		</dependency>
 	</dependencies>
 </project>
